updated script for processing emails into a csv; new output uploaded
Completely revamped script for parsing the emails, now uses python's email and
csv modules for reading and writing. A description of the csv output is in the
comment at the head of the process_data.py script.

The csv is in output.txt.gz (~670MB unzipped!). output.log.gz is a log of the
run -- there are about 18,000 emails not included because they did not include
To,Cc or Bcc fields; some (9433) had X-To fields, others (8650) didn't even
have those...